scripts/requirements.txt
requests>=2.28.0
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
